+++ lib/network.go
@@ -6783,7 +6783,7 @@ type UserAssociationQuery struct {
 	AssociationTypePrefix  string
 	AssociationValue       string
 	AssociationValuePrefix string
-	Limit                  uint64
+	Limit                  int
 	LastSeenAssociationID  *BlockHash
 	SortDescending         bool
 }
@@ -6796,7 +6796,7 @@ type PostAssociationQuery struct {
 	AssociationTypePrefix  string
 	AssociationValue       string
 	AssociationValuePrefix string
-	Limit                  uint64
+	Limit                  int
 	LastSeenAssociationID  *BlockHash
 	SortDescending         bool
 }
